requirements.txt
# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
python-multipart>=0.0.5

# CLI
click>=8.0.0

# Subtitle Processing
pysrt>=1.1.2
webvtt-py>=0.4.6
nltk>=3.6.0
language-tool-python>=2.7.1

# Video Processing
opencv-python>=4.5.0
ffmpeg-python>=0.2.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utils
python-dotenv>=0.19.0
pydantic>=1.9.0
requests>=2.26.0
